Created mixin methods for text selection based logic adapters.
